poll_wait is now float to allow for shorter poll intervals than full seconds, e.g. 0.5 for 500msec

+++ image-segmentation-keras/1.14.0_0.3.0/predict.py
@@ -159,7 +159,7 @@ def predict_on_images(model, input_dir, output_dir, tmp_dir, delete_input,
     :param colors: the list of colors to use (flat list of r,g,b values, eg "0, 0, 0, 127, 127, 127"), will get padded to 256 triplets
     :type colors: list
     :param poll_wait: the amount of seconds between polls when not in watchdog mode
-    :type poll_wait: int
+    :type poll_wait: float
     :param continuous: whether to poll for files continuously
     :type continuous: bool
     :param use_watchdog: whether to react to file creation events rather than use fixed-interval polling
@@ -206,7 +206,7 @@ if __name__ == '__main__':
     parser.add_argument('--prediction_in', help='Path to the test images', required=True, default=None)
     parser.add_argument('--prediction_out', help='Path to the output csv files folder', required=True, default=None)
     parser.add_argument('--prediction_tmp', help='Path to the temporary csv files folder', required=False, default=None)
-    parser.add_argument('--poll_wait', type=int, help='poll interval in seconds when not using watchdog mode', required=False, default=1)
+    parser.add_argument('--poll_wait', type=float, help='poll interval in seconds when not using watchdog mode', required=False, default=1.0)
     parser.add_argument('--continuous', action='store_true', help='Whether to continuously load test images and perform prediction', required=False, default=False)
     parser.add_argument('--use_watchdog', action='store_true', help='Whether to react to file creation events rather than performing fixed-interval polling', required=False, default=False)
     parser.add_argument('--watchdog_check_interval', type=int, help='check interval in seconds for the watchdog', required=False, default=10)
